Added a appsetting to recreate the db or not

+++ SchoolMachine.API/Extensions/ApplicationBuilderExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using SchoolMachine.DataAccess.Entities;
 
@@ -6,13 +7,18 @@ namespace SchoolMachine.API.Extensions
 {
     public static class ApplicationBuilderExtensions
     {
-        public static void DeleteAndRecreateDatabase(this IApplicationBuilder app)
+        public static void DeleteAndRecreateDatabase(this IApplicationBuilder app, IConfiguration configuration)
         {
-            using (var serviceScope = app.ApplicationServices.GetService<IServiceScopeFactory>().CreateScope())
+            var appSettingsSection = configuration.GetSection("AppSettings");
+            var isRecreateDatabase = appSettingsSection["IsRecreateDatabase"];
+            if (isRecreateDatabase.ToLower() == "true")
             {
-                var context = serviceScope.ServiceProvider.GetRequiredService<SchoolMachineContext>();
-                context.Database.EnsureDeleted();
-                context.Database.EnsureCreated();
+                using (var serviceScope = app.ApplicationServices.GetService<IServiceScopeFactory>().CreateScope())
+                {
+                    var context = serviceScope.ServiceProvider.GetRequiredService<SchoolMachineContext>();
+                    context.Database.EnsureDeleted();
+                    context.Database.EnsureCreated();
+                }
             }
         }
     }
